Move scripts into packages/types/src/scripts
