+++ package.json
@@ -15,8 +15,8 @@
   "scripts": {
     "build": "yarn run build:methodsdoc && polkadot-dev-build-ts && polkadot-dev-build-docs",
     "build:htmldoc": "yarn clean && typedoc --theme default --out docs/html",
-    "build:methodsdoc": "node packages/types/src/MetadataMdWrapper.js",
-    "build:decodedjson": "node packages/types/src/MetadataJsonDecodeWrapper.js",
+    "build:methodsdoc": "node packages/types/src/scripts/MetadataMdWrapper.js",
+    "build:decodedjson": "node packages/types/src/scripts/MetadataJsonDecodeWrapper.js",
     "check": "tslint --project . && tsc --noEmit",
     "clean": "polkadot-dev-clean-build",
     "postinstall": "polkadot-dev-yarn-only",
