+++ packages/types/src/scripts/MetadataJsonDecodeWrapper.js
@@ -3,5 +3,5 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-require('../babelConfig.js');
-require('./MetadataMd.ts');
+require('../../babelConfig.js');
+require('./MetadataJsonDecode.ts');
